# specs/027-notifications/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Market Lens Notifications API
  version: 1.0.0
  description: >
    What a person has asked to be told about, where, and when not to be. Every operation is private
    to the authenticated caller, except the unsubscribe path, which is authorised by a signed link
    and can change exactly one preference.

    Nothing here sends anything a person did not ask for, and nothing is on by default.
servers:
  - url: /api/v1
x-market-lens-access-boundary:
  user_private:
    - /notifications/preferences
    - /notifications/quiet-hours
    - /notifications/subscriptions
    - /notifications/subscriptions/{id}
    - /notifications/history
    - /notifications/test-email
    - /notifications/push-key
  public:
    - /notifications/unsubscribe
  rule: >
    Every private operation requires an active session, is scoped to the caller's own account, and
    is refused to a deactivated account. The unsubscribe path takes a signed, expiring,
    single-purpose token instead, and can only disable one kind on one channel.

paths:
  /notifications/preferences:
    get:
      summary: What this person has asked to be told about
      responses:
        "200":
          description: Every kind and channel available to this person, with its state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Preferences" }
        "401": { description: No active session }
    put:
      summary: Turn one kind on or off for one channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, channel, enabled]
              properties:
                kind: { $ref: "#/components/schemas/Kind" }
                channel: { $ref: "#/components/schemas/Channel" }
                enabled: { type: boolean }
      responses:
        "200":
          description: Every preference, after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Preferences" }
        "400": { $ref: "#/components/responses/Refusal" }
        "401": { description: No active session }

  /notifications/quiet-hours:
    put:
      summary: Set or clear the hours during which nothing arrives
      description: >
        A window whose end is before its start crosses midnight. A notification raised inside the
        window is held until it ends — never sent early, never dropped. Sending null clears it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                starts_at: { type: [string, "null"], examples: ["22:00"] }
                ends_at: { type: [string, "null"], examples: ["07:00"] }
                timezone: { type: [string, "null"], examples: ["Europe/Stockholm"] }
      responses:
        "200":
          description: Every preference, after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Preferences" }
        "400": { $ref: "#/components/responses/Refusal" }
        "401": { description: No active session }

  /notifications/push-key:
    get:
      summary: The instance's public push key
      description: >
        Needed by the browser to create a subscription. The key pair is generated by this product on
        first start and kept in its own database, so a restored backup keeps working. Only the
        public half is ever served.
      responses:
        "200":
          description: The public key, base64url, unpadded
          content:
            application/json:
              schema:
                type: object
                required: [public_key]
                properties:
                  public_key: { type: string }
        "401": { description: No active session }

  /notifications/subscriptions:
    get:
      summary: The devices this person has subscribed
      responses:
        "200":
          description: One entry per device
          content:
            application/json:
              schema:
                type: object
                required: [subscriptions]
                properties:
                  subscriptions:
                    type: array
                    items: { $ref: "#/components/schemas/Subscription" }
        "401": { description: No active session }
    post:
      summary: Subscribe this device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, p256dh, auth, label]
              description: >
                Exactly what the browser produced, plus a label the person will recognise. Nothing
                is recorded about the device itself — a user agent is a fingerprint and is not
                needed to send a push.
              properties:
                endpoint: { type: string }
                p256dh: { type: string, description: base64url, unpadded }
                auth: { type: string, description: base64url, unpadded }
                label: { type: string, maxLength: 120 }
      responses:
        "201":
          description: The devices, after subscribing
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items: { $ref: "#/components/schemas/Subscription" }
        "400": { $ref: "#/components/responses/Refusal" }
        "401": { description: No active session }

  /notifications/subscriptions/{id}:
    delete:
      summary: Stop sending to this device
      description: The subscription is deleted rather than muted, from any signed-in device.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: The devices, after revoking }
        "401": { description: No active session }
        "404": { description: No such device for this person }

  /notifications/history:
    get:
      summary: What was sent, when, and what failed
      description: >
        Without the payload. What a message said is not kept beyond what was needed to send it.
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                type: object
                required: [notifications]
                properties:
                  notifications:
                    type: array
                    items: { $ref: "#/components/schemas/NotificationRecord" }
        "401": { description: No active session }

  /notifications/test-email:
    post:
      summary: Prove a message actually arrives
      description: >
        Sends one short message to the caller's own address. The settings check proves the server
        accepts a connection; this proves a message arrives, which is a different failure — a server
        can connect, refuse the sender, and look configured. It takes no recipient: one would be a
        way to make this installation mail a stranger.
      responses:
        "200":
          description: The message was accepted by the mail server
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: boolean }
        "400": { $ref: "#/components/responses/Refusal" }
        "401": { description: No active session }

  /notifications/unsubscribe:
    post:
      summary: Turn off one kind on one channel, from a link in an email
      description: >
        Authorised by a signed, expiring token naming one person, one kind and one channel. It can
        do nothing else: it cannot read, cannot enable, and cannot change a second preference. An
        unsubscribe that required signing in is one people do not use — they mark the mail as spam
        instead.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: What was turned off
          content:
            application/json:
              schema:
                type: object
                required: [kind, channel, stopped]
                properties:
                  kind: { $ref: "#/components/schemas/Kind" }
                  channel: { $ref: "#/components/schemas/Channel" }
                  stopped: { type: boolean }
        "400": { description: The link is not readable, has expired, or has already been used }

components:
  responses:
    Refusal:
      description: Refused, with a reason a person can act on
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Refusal" }

  schemas:
    Refusal:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [unknown_kind, unknown_channel, not_available_to_you, invalid_quiet_hours,
                     unknown_timezone, invalid_subscription, invalid_token]
            message: { type: string }

    Kind:
      type: string
      enum: [decision_waiting, paper_fill, pipeline_failure, signal_change]

    Channel:
      type: string
      enum: [email, web_push]

    Preference:
      type: object
      required: [kind, channel, enabled]
      properties:
        kind: { $ref: "#/components/schemas/Kind" }
        channel: { $ref: "#/components/schemas/Channel" }
        enabled: { type: boolean }

    Preferences:
      type: object
      required: [preferences, quiet_hours, nothing_is_on_by_default]
      properties:
        preferences:
          type: array
          description: >
            Only the kinds available to this person. pipeline_failure is absent for a member rather
            than present and refused — it is the one kind nobody but the owner can act on.
          items: { $ref: "#/components/schemas/Preference" }
        quiet_hours:
          type: [object, "null"]
          properties:
            starts_at: { type: string }
            ends_at: { type: string }
            timezone: { type: string }
        nothing_is_on_by_default:
          type: boolean
          description: >
            Always true and always present. Every kind is off until the person asks for it; a
            product that mails somebody unasked has decided on their behalf.

    Subscription:
      type: object
      required: [id, label, created_at]
      description: >
        No endpoint and no keys are returned. They are what is needed to send to the device, not
        something a page has any use for.
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        created_at: { type: string, format: date-time }
        last_used_at: { type: [string, "null"], format: date-time }

    NotificationRecord:
      type: object
      required: [kind, channel, state, created_at]
      properties:
        kind: { $ref: "#/components/schemas/Kind" }
        channel: { $ref: "#/components/schemas/Channel" }
        state: { type: string, enum: [pending, sent, failed, abandoned] }
        count: { type: integer }
        attempts: { type: integer }
        last_error: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        sent_at: { type: [string, "null"], format: date-time }